Fix PHPStan CI failures from larastan inference drift
CI installs fresh dependencies, and the newer larastan infers
Collection<(int|string), ...> keys for Model::find(mixed), so the
baselined messages no longer matched and unmatched ignores errored.

- Cast the scalar ids passed to TransactionType::find and
  Community::find, removing the Collection|Model unions.
- Set reportUnmatchedIgnoredErrors false (matching the other
  packages) so baseline drift cannot fail the build by itself.
- Regenerate the baseline against current stable deps.

+++ src/Http/Middleware/EnsureIsCommunityOwner.php
@@ -16,7 +16,7 @@ class EnsureIsCommunityOwner
      */
     public function handle(Request $request, Closure $next)
     {
-        $community = Community::find(Auth::user()->currentCommunity->id);
+        $community = Community::find((int) Auth::user()->currentCommunity->id);
 
         if (Auth::user()->id != $community->user_id) {
             abort(403, 'Unauthorized action');
